Refactor to use BigInteger

+++ src/RunnerInterface.java
@@ -1,11 +1,13 @@
+import java.math.BigInteger;
+
 public interface RunnerInterface {
     /**
-     * For the secret sharing protocol, the Runner class has the secret. This secret is
-     * taken from the user if specified, or defaulted to 50.
+     * For the secret sharing protocol, the Runner class has the secret. Ideally, this
+     * secret should be taken from the user, but currently defaults to 50.
      *
      * @return the secret.
      */
-    int getSecretFromUser();
+    BigInteger getSecretFromUser();
 
     /**
      * Calcluates the value of f(i) for i in range 0 to n, where n is the number of peers.
@@ -13,7 +15,7 @@ public interface RunnerInterface {
      * @param polynomial the current Polynomial object.
      * @return an array such that f[i] = f(i), and f is the current polynomial.
      */
-    int[] getF(Polynomial polynomial);
+    BigInteger[] getF(Polynomial polynomial);
 
     /**
      * Sends the shares to the corresponding peer. share i is sent to peer i, along
@@ -21,5 +23,5 @@ public interface RunnerInterface {
      * @param f an array such that f[i] = f(i). See {@code getSecretFromUser()}
      * @param serviceName the host name of the service. Used for InetAddress.getByName().
      */
-    void distributeShares(int[] f, String serviceName);
+    void distributeShares(BigInteger[] f, String serviceName);
 }
